refactor: stores를 src/로 이동
  - stores/ → src/stores/
