+++ src/stores/notification/types.ts
