+++ src/stores/onboarding/types.ts
